Add HHL_experiment.py and HHL_plots_experimet.py
HHL_experiment.py : Contains the simulation of the HHL algorithm. Then generate random systems solve them using the original algorithm and then also performing the post selection on the c-register. It calculates the error compearing the output in both cases with the exact solution. It saves all the results on a folder call data_experiment

HHL_plots_experimet.py : Read the data saved by the previous file and plot it.

This are the final version of the files, they should not be modifyed